Rework layout items
Update LayItem trait, which now has methods to retrieve a BaseItem, on
which the library will work.

This allows the library to work on different kind of objects, without
having all classes implement all the getters/setters for the fields of
the BaseItem.

The LayoutItem trait now also has functions to handle mouse clicks and
key presses (triggered by BaseItem.click() / BaseItem.key()).

Signed-off-by: Paul Cercueil <paul@crapouillou.net>